Aria project isolated and pa11y config file split
