+++ .dockerignore
@@ -2,7 +2,9 @@
 .git
 .sass-cache
 build
-node_modules/
+/node_modules/
+/test/aria/node_modules/
+/test/e2e/node_modules/
 coverage/
 coverage-legacy/
 dist/
