+++ Dockerfile
@@ -12,9 +12,8 @@ RUN apt-get update && \
 
 COPY package.json package-lock.json /app/
 
-# Install all NPM dependencies. Also:
+# Install all NPM dependencies, and:
 #  * Changing git URL because network is blocking git protocol...
-#  * Using `CYPRESS_INSTALL_BINARY` to skip installing Cypress for it is not used here
 RUN git config --global url."https://".insteadOf git:// && \
     git config --global url."https://github.com/".insteadOf git@github.com: && \
     npm config set registry https://repo.datapunt.amsterdam.nl/repository/npm-group/ && \
