+++ docker-compose.yml
@@ -6,6 +6,8 @@ services:
         - BUILD_ENV=acc
       context: .
     ipc: host
+    ports:
+    - 8080:80
 
   test-unit-integration:
     build:
@@ -33,8 +35,8 @@ services:
 
   test-e2e-aria:
     build:
-      dockerfile: Dockerfile-e2e-aria
-      context: .
+      dockerfile: Dockerfile
+      context: ./test/aria/
     links:
       - atlas:atlas
     environment:
@@ -43,7 +45,7 @@ services:
       - USERNAME_EMPLOYEE_PLUS
       - PASSWORD_EMPLOYEE
       - PASSWORD_EMPLOYEE_PLUS
-    command: ./scripts/ci/test-e2e-aria.sh
+    command: ./test-e2e-aria.sh
 
   test-health-checks:
     image: build.datapunt.amsterdam.nl:5000/datapunt/atlas-health-checks:latest
