--- pa11y/.gitignore
@@ -1,1 +1,0 @@
-*.png
