+++ test/aria/.babelrc
@@ -1,4 +1,4 @@
 {
-  "presets": ["env"],
-  "plugins": ["transform-object-rest-spread", "transform-class-properties"]
+    "presets": ["react-app", "env"],
+    "plugins": ["transform-object-rest-spread", "transform-class-properties"]
 }
