+++ test/aria/.gitignore
@@ -1,3 +1,2 @@
 .env
-.git
 /node_modules/
