+++ test/aria/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "atlas-e2e-aria",
-  "version": "1.0.0",
+  "name": "atlas",
+  "version": "2.1.1",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Amsterdam/atlas.git"
@@ -11,10 +11,85 @@
   },
   "author": "Datapunt Amsterdam",
   "license": "MPL-2.0",
+  "pre-commit": "lint",
   "scripts": {
-    "test-e2e-aria": "babel-node ./pa11y.js"
+    "build": "npm run build-prod",
+    "build-acc": "grunt && webpack --config webpack.acc.js",
+    "build-dev": "grunt && webpack --config webpack.dev.js",
+    "build-prod": "grunt && webpack --config webpack.prod.js",
+    "build-test": "grunt && webpack --config webpack.acc.js",
+    "clean": "npm run clean-cache && npm run clean-dirs",
+    "clean-cache": "npm cache verify",
+    "clean-dirs": "rimraf ./node_modules ./coverage ./coverage-legacy ./dist ./build ./pa11y/*.png",
+    "lint": "npm run lint:js && npm run lint:style",
+    "lint:js": "eslint --format=node_modules/eslint-formatter-pretty src modules test/cypress --ext js,jsx",
+    "lint:style": "sass-lint -c src/.sass-lint.yml 'src/**/*.scss' --verbose --no-exit && sass-lint -c modules/.sass-lint.yml 'modules/**/*.scss' --verbose --no-exit",
+    "preinstall": "npm cache verify",
+    "start": "grunt && webpack-dev-server --config webpack.dev.js",
+    "start-prod": "grunt && webpack-dev-server --config webpack.prod-api-dev.js",
+    "test": "npm run lint && cross-env karma start && jest --config=./test/jest.unit.config.js",
+    "test-e2e-aria": "babel-node ./test/pa11y.js",
+    "test-e2e": "cross-env API_ROOT=https://acc.api.data.amsterdam.nl node ./scripts/cypress/cypress-run.js",
+    "test-e2e-prod": "cross-env API_ROOT=https://api.data.amsterdam.nl node ./scripts/cypress/cypress-run.js",
+    "test-e2e-prod-watch": "cross-env API_ROOT=https://api.data.amsterdam.nl node ./scripts/cypress/cypress-watch.js",
+    "test-e2e-watch": "cross-env API_ROOT=https://acc.api.data.amsterdam.nl node ./scripts/cypress/cypress-watch.js",
+    "test-eslint": "eslint ./modules/ ./src/ ./test/cypress/",
+    "test-integration": "cross-env NODE_ENV=development jest --config=./test/jest.integration.config.js",
+    "test-integration-watch": "cross-env NODE_ENV=development jest --config=./test/jest.integration.config.js --watch",
+    "test-jest": "cross-env NODE_ENV=development jest --config=./test/jest.unit.config.js",
+    "test-jest-watch": "cross-env NODE_ENV=development jest --config=./test/jest.unit.config.js --watch --onlyChanged --coverage",
+    "test-karma": "karma start",
+    "test-karma-watch": "karma start --no-single-run",
+    "dyson:api": "nodemon --watch test/dyson/mock/api --exec babel-node --presets=latest test/dyson/mock/api"
   },
   "dependencies": {
+    "angular": "~1.5.11",
+    "angular-aria": "~1.5.11",
+    "angular-i18n": "~1.5.11",
+    "angular-sanitize": "~1.5.11",
+    "angulartics-piwik": "^1.0.6",
+    "babel-preset-react-app": "^3.1.2",
+    "bbga_visualisatie_d3": "https://github.com/Amsterdam/bbga_visualisatie_d3.git#master",
+    "cross-env": "^5.2.0",
+    "d3": "^3.5.17",
+    "deep-copy": "1.4.2",
+    "escape-string-regexp": "1.0.5",
+    "history": "4.7.2",
+    "isomorphic-fetch": "2.2.1",
+    "leaflet": "^1.3.4",
+    "leaflet-draw": "^1.0.0",
+    "leaflet-rotatedmarker": "^0.2.0",
+    "leaflet.markercluster": "^1.4.1",
+    "leaflet.nontiledlayer": "^1.0.7",
+    "lodash.debounce": "^4.0.8",
+    "lodash.defer": "^4.1.0",
+    "lodash.get": "^4.4.2",
+    "lodash.has": "^4.5.2",
+    "lodash.identity": "^3.0.0",
+    "lodash.isequal": "^4.5.0",
+    "marked": "^0.5.0",
+    "marzipano": "https://github.com/Amsterdam/marzipano.git#master",
+    "objectFitPolyfill": "^2.1.1",
+    "proj4": "^2.4.4",
+    "proj4leaflet": "^1.0.2",
+    "prop-types": "^15.6.0",
+    "raven-js": "^3.19.1",
+    "react": "16.2.0",
+    "react-dom": "^16.2.0",
+    "react-leaflet": "^1.9.0",
+    "react-leaflet-markercluster": "^1.1.8",
+    "react-redux": "^5.0.7",
+    "react-resize-aware": "^2.7.0",
+    "recharts": "^1.2.0",
+    "redux": "^3.7.2",
+    "redux-saga": "^0.16.0",
+    "remove-markdown": "^0.2.2",
+    "reselect": "^3.0.1"
+  },
+  "devDependencies": {
+    "@metahub/karma-jasmine-jquery": "^2.0.0",
+    "angular-mocks": "~1.5.11",
+    "autoprefixer": "^9.1.5",
     "babel-cli": "^6.26.0",
     "babel-jest": "^22.0.4",
     "babel-loader": "^7.1.2",
@@ -24,8 +99,69 @@
     "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.7.0",
     "chalk": "2.3.0",
+    "clean-webpack-plugin": "^0.1.19",
+    "copy-webpack-plugin": "^4.5.2",
+    "css-loader": "^1.0.0",
+    "dyson": "^2.0.1",
+    "dyson-generators": "^0.2.0",
+    "dyson-image": "^0.2.0",
+    "enzyme": "^3.5.1",
+    "enzyme-adapter-react-16": "^1.4.0",
+    "enzyme-to-json": "^3.1.4",
+    "eslint": "^3.19.0",
+    "eslint-config-airbnb": "^15.1.0",
+    "eslint-config-angular": "^0.5.0",
+    "eslint-config-standard": "^6.2.1",
+    "eslint-formatter-pretty": "^1.3.0",
+    "eslint-plugin-angular": "^1.4.1",
+    "eslint-plugin-cypress": "^2.0.1",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jasmine": "^2.9.1",
+    "eslint-plugin-jsx-a11y": "^5.1.1",
+    "eslint-plugin-promise": "^3.6.0",
+    "eslint-plugin-react": "^7.4.0",
+    "eslint-plugin-standard": "^2.0.1",
+    "file-loader": "^2.0.0",
+    "grunt": "^1.0.3",
+    "grunt-svg-sprite": "^1.3.6",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "html-webpack-template": "^6.2.0",
+    "identity-obj-proxy": "^3.0.0",
+    "imports-loader": "^0.7.1",
+    "jasmine-core": "^2.99.0",
+    "jest": "^22.0.4",
+    "jest-environment-jsdom": "^22.4.0",
+    "jest-environment-jsdom-global": "^1.0.3",
+    "jest-environment-node": "^22.0.4",
+    "jest-fetch-mock": "^1.3.3",
+    "karma": "1.7.1",
+    "karma-babel-preprocessor": "^7.0.0",
+    "karma-coverage": "^1.1.2",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
+    "karma-jasmine": "^1.0.2",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-phantomjs-launcher": "^1.0.0",
+    "karma-sourcemap-loader": "^0.3.7",
+    "karma-webpack": "^3.0.5",
+    "mini-css-extract-plugin": "^0.4.2",
+    "node-sass": "^4.9.3",
     "p-all": "1.0.0",
     "pa11y": "5.0.4",
-    "pa11y-reporter-cli": "^1.0.0"
+    "pa11y-reporter-cli": "^1.0.0",
+    "phantomjs-prebuilt": "^2.1.8",
+    "postcss-loader": "^3.0.0",
+    "pre-commit": "^1.2.2",
+    "redux-mock-store": "^1.3.0",
+    "redux-saga-test-plan": "^3.6.0",
+    "rimraf": "^2.6.2",
+    "sass-lint": "1.12.1",
+    "sass-loader": "^7.1.0",
+    "svg-react-loader": "^0.4.5",
+    "uglifyjs-webpack-plugin": "^2.0.0",
+    "webpack": "^4.19.0",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-server": "^3.1.8",
+    "webpack-merge": "^4.1.4"
   }
 }
